+++ CHANGELOG.md
@@ -20,6 +20,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 #### Router
 * `Linna\Router\Router->__construct()` now require a `Linna\Router\RouteCollection` as first argument
 
+#### Storage
+* `Linna\Storage\ConnectorsInterface->getResource()` now has return type `object`
+* `Linna\Storage\Connectors\PdoConnector->getResource()` now has return type `object`
+* `Linna\Storage\Connectors\MysqliConnector->getResource()` now has return type `object`
+* `Linna\Storage\Connectors\MongoDBConnector->getResource()` now has return type `object`
+
 ## [v0.25.0](https://github.com/linna/framework/compare/v0.24.0...v0.25.0) - 2019-01-13
 
 ### Added